# sportrait-import-swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Initial Sportrait Import API description - an API to process events and\nphoto imports on the Sportrait Server\n",
    "version" : "0.0.1",
    "title" : "Sportrait Import API",
    "termsOfService" : "https://sportrait.com/terms/",
    "contact" : {
      "email" : "info@sportrait.com",
      "name" : "Alexander Josef"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/events/{eventId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "event" ],
        "summary" : "return an existing event",
        "operationId" : "getEvent",
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Event not found"
          }
        }
      },
      "put" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "event" ],
        "summary" : "Update an existing event",
        "operationId" : "updateEvent",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Event"
        },
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Event not found"
          },
          "405" : {
            "description" : "Validation exception"
          }
        }
      },
      "delete" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "description" : "ID of the event to delete",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "event" ],
        "summary" : "delete an event incl. all of its eventCategories, albums & photos - TODO add security",
        "operationId" : "deleteEvent",
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "403" : {
            "description" : "Forbidden - Not authorized to delete ressource"
          },
          "404" : {
            "description" : "Not found"
          },
          "405" : {
            "description" : "Given eventId doesn't identify an event ressource"
          }
        }
      }
    },
    "/events/{eventId}/eventCategories/" : {
      "post" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "description" : "ID of the event for the eventCategory to be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "event" ],
        "summary" : "Add an eventCategory",
        "operationId" : "addEventCategory",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/EventCategory"
        },
        "responses" : {
          "201" : {
            "description" : "successful operation - added",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventCategory"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/eventCategories/{eventCategoryId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "eventCategoryId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "eventCategory" ],
        "summary" : "Return an eventCategory identified by eventCategoryId",
        "operationId" : "getEventCategory",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventCategory"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Not found"
          },
          "403" : {
            "description" : "forbidden"
          }
        }
      },
      "put" : {
        "parameters" : [ {
          "name" : "eventCategoryId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "eventCategory" ],
        "summary" : "Update an eventCategory",
        "operationId" : "updateEventCategory",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/EventCategory"
        },
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventCategory"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      },
      "delete" : {
        "parameters" : [ {
          "name" : "eventCategoryId",
          "in" : "path",
          "description" : "ID of the eventCategory to delete",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "eventCategory" ],
        "summary" : "delete an eventCategory incl. all of its albums & photos - TODO add security",
        "operationId" : "deleteEventCategory",
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      }
    },
    "/eventCategories/{eventCategoryId}/albums" : {
      "get" : {
        "parameters" : [ {
          "name" : "eventCategoryId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "eventCategory" ],
        "summary" : "Get all albums for given eventCategory.",
        "operationId" : "listAlbumsForEventCategory",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Album"
                  }
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      },
      "post" : {
        "parameters" : [ {
          "name" : "eventCategoryId",
          "in" : "path",
          "description" : "ID of the eventCategory where the album will be added",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "tags" : [ "album" ],
        "summary" : "Add (create) an album for a given eventCategory. Uploads associated photos to the album. An existing album for the given eventCategory can be used on the backend (photos will be added) if one exist.\n  Todo, add query parameters for apply number recognition etcetc.\n",
        "operationId" : "addAlbum",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Album"
        },
        "responses" : {
          "201" : {
            "description" : "successful operation - album created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Album"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      }
    },
...
  },
...
  "servers" : [ {
    "url" : "http://localhost:8080/api/import"
  }, {
    "url" : "https://sportrait.com/api/import/1.0.0"
  }, {
    "url" : "https://int2.sportrait.com/api/import/1.0.0"
  } ],
  "components" : {
    "requestBodies" : {
      "Event" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Event"
            }
          },
          "application/xml" : {
            "schema" : {
              "$ref" : "#/components/schemas/Event"
            }
          }
        },
        "description" : "Event object that needs to be added to the server",
        "required" : true
      },
...
      "EventCategory" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/EventCategory"
            }
          }
        },
        "required" : true
      },
      "Album" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Album"
            }
          }
        },
        "required" : true
      },
...
    },
    "schemas" : {
      "Event" : {
        "type" : "object",
        "required" : [ "title", "navTitle", "date", "organizerUrl" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string",
            "example" : "sola 2020"
          },
          "navTitle" : {
            "type" : "string",
            "example" : "sola 2020"
          },
          "organizerUrl" : {
            "type" : "string",
            "format" : "uri",
            "example" : "https://asvz.ch/sola"
          },
          "date" : {
            "type" : "string",
            "format" : "date",
            "example" : "2020-12-13",
            "description" : "The date (day) the event takes, or took, place, in the format YYYY-MM-DD."
          },
          "zipCode" : {
            "type" : "string",
            "example" : "8000"
          },
          "city" : {
            "type" : "string",
            "example" : "Zurich"
          },
          "description" : {
            "type" : "string",
            "example" : "Die 20182-te Durchführung der Sola führt rund um Zürich"
          },
          "eventCategories" : {
            "type" : "array",
            "xml" : {
              "name" : "eventCategories",
              "wrapped" : true
            },
            "items" : {
              "$ref" : "#/components/schemas/EventCategory"
            }
          },
          "status" : {
            "type" : "string",
            "description" : "event status",
            "enum" : [ "new", "mapped", "online" ]
          }
        }
      },
      "Album" : {
        "type" : "object",
        "required" : [ "title", "photosS3Uri" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string",
            "example" : "sola 2020 - etappe 12"
          },
          "navTitle" : {
            "type" : "string",
            "description" : "the slug of the album",
            "example" : "sola-2020-etappe-12"
          },
          "photosCount" : {
            "type" : "integer",
            "description" : "the total number of photos in the album (might change during import)",
            "example" : 3741
          },
          "description" : {
            "type" : "string",
            "example" : "die 12. etappe der sola 2030"
          },
          "freeHighresDownload" : {
            "type" : "boolean",
            "description" : "whether this album allows a free high resulation download of the digital image.",
            "example" : true,
            "default" : false
          },
          "applyLogoOnFineImages" : {
            "type" : "boolean",
            "description" : "whether to process the fine images with a logo or watermark. Default to false with the newer imgix based import.",
            "example" : true,
            "default" : false
          },
          "photosS3Uri" : {
            "type" : "string",
            "description" : "location to find the images to be processed",
            "example" : "s3://dev.ireland.photos.sportrait.com/upload/2018-etappe-1-rekognition-test/"
          },
          "sponsorBarRelativeUrl" : {
            "type" : "string",
            "description" : "relative URL (without bucket) to sponsor bar - not used yet",
            "example" : "logo/2020/asvz-logo-2020.png"
          },
          "asvzLogoRelativeUrl" : {
            "type" : "string",
            "description" : "relative URL (without bucket) to logo - not used yet",
            "example" : "logo/2020/asvz-logo-2020.png"
          },
          "creationDate" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Creation Date / Time for this album as defined by RFC 3339, section 5.6, for example, 2017-07-21T17:32:28Z\nCAUTION: this information is not yet available - a date shown here is the upload date of the first photo in the album\n",
            "example" : "2017-07-21T17:32:28Z"
          },
          "status" : {
            "type" : "string",
            "description" : "album publication status",
            "enum" : [ "hidden", "published", "importing" ]
          }
        }
      },
      "EventCategory" : {
        "type" : "object",
        "required" : [ "title", "description", "status" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string",
            "example" : "sola 2020 - etappe 12"
          },
          "description" : {
            "type" : "string",
            "example" : "die 12. etappe der sola 2030"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "the relative position / order of this eventCategory within the event",
            "example" : 0
          },
          "status" : {
            "type" : "string",
            "description" : "event status",
            "enum" : [ "mapped", "published", "emtpy" ]
          }
        }
      },
...
    }
  }
}